Create Acreage_Update.py
Code used to update lot_size_sq_footage attribute in the database to the lot's size in acres, rounded down.  We are getting a NaN/Infinity error when trying to run the sold_classifier RF on the server. The goal was to prevent these errors by using the smaller acreage value for the lots.